+++ .storybook/preview.js
@@ -72,6 +72,7 @@ const fontPathMap = [
   { prefix: 'F_MALLANNA', path: 'fonts/Mallanna/' },
   { prefix: 'F_NOTO_SANS_ETHIOPIC', path: 'fonts/NotoSansEthiopic/' },
   { prefix: 'F_NOTO_SERIF_BENGALI', path: 'fonts/NotoSerifBengali/' },
+  { prefix: 'F_NOTO_SERIF_SINHALA', path: 'fonts/NotoSerifSinhala/' },
   { prefix: 'F_PADAUK', path: 'fonts/Padauk/' },
   { prefix: 'F_REITH_QALAM', path: 'fonts/ReithQalam/' },
   { prefix: 'F_REITH_SANS', path: 'fonts/Reith/' },