+++ packages/utilities/gel-foundations/index.test.jsx
@@ -92,6 +92,7 @@ const scriptsExpectedExports = {
   nepali: 'object',
   noAscendersOrDescenders: 'object',
   sinhalese: 'object',
+  sinhaleseNotoSerif: 'object',
   tamil: 'object',
   thai: 'object',
 };