+++ packages/utilities/psammead-styles/index.test.jsx
@@ -27,6 +27,8 @@ const fontsExpectedExports = {
   F_REITH_QALAM_BOLD: 'function',
   F_ISKOOLA_POTA_BBC_REGULAR: 'function',
   F_ISKOOLA_POTA_BBC_BOLD: 'function',
+  F_NOTO_SERIF_SINHALA_REGULAR: 'function',
+  F_NOTO_SERIF_SINHALA_BOLD: 'function',
   F_LATHA_REGULAR: 'function',
   F_LATHA_BOLD: 'function',
   F_TAMIL_REGULAR: 'function',