+++ packages/components/psammead-section-label/CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 7.1.4 | [PR#4548](https://github.com/bbc/psammead/pull/4548) Bumps dependencies |
 | 7.1.3 | [PR#4547](https://github.com/bbc/psammead/pull/4547) Bumps packages |
 | 7.1.2 | [PR#4545](https://github.com/bbc/psammead/pull/4545) Bump dependencies for @bbc/psammead-styles |
 | 7.1.1 | [PR#4552](https://github.com/bbc/psammead/pull/4552) Bump psammead-styles |