+++ packages/components/psammead-timestamp/CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 4.0.23 | [PR#4548](https://github.com/bbc/psammead/pull/4548) Bumps dependencies |
 | 4.0.22 | [PR#4547](https://github.com/bbc/psammead/pull/4547) Bumps packages |
 | 4.0.21 | [PR#4545](https://github.com/bbc/psammead/pull/4545) Bump dependencies for @bbc/psammead-styles |
 | 4.0.20 | [PR#4552](https://github.com/bbc/psammead/pull/4552) Bump psammead-styles |